Chore: add assign therapist get route

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\AdminDashboardController;
+use App\Http\Controllers\AdminTherapistController;
 use App\Http\Controllers\JournalController;
 use App\Http\Controllers\RegisteredUserController;
 use App\Http\Controllers\RegisteredTherapistController;
@@ -17,8 +18,8 @@ Route::get('/', function () {
 Route::get('/register', [RegisteredUserController::class, 'create']);
 Route::post('/register', [RegisteredUserController::class, 'store'])->name('register.user');
 
-Route::get('/register-therapist', [RegisteredTherapistController::class, 'create']);
-Route::post('/register-therapist', [RegisteredTherapistController::class, 'store'])->name('register.therapist');
+Route::get('/register-therapist', [RegisteredTherapistController::class, 'create'])->middleware(['auth', Role::class . ':admin']);
+Route::post('/register-therapist', [RegisteredTherapistController::class, 'store'])->name('register.therapist')->middleware(['auth', Role::class . ':admin']);
 
 
 Route::get('/login', [SessionUserController::class, 'create'])->name('login');
@@ -26,13 +27,16 @@ Route::post('/login', [SessionUserController::class, 'store']);
 
 // Dashboards
 Route::get('/dashboard/user', [UserDashboardController::class, 'index'])->middleware(['auth', Role::class . ':user']);
-Route::get('/dashboard/admin', [AdminDashboardController::class, 'index'])->middleware('auth');
+Route::get('/dashboard/admin', [AdminDashboardController::class, 'index'])->middleware(['auth', Role::class . ':admin']);
 
 // User Journal
-Route::get('/journals', [JournalController::class, 'index'])->middleware('auth');
-Route::get('/journals/{journal}', [JournalController::class, 'show'])->middleware('auth');
+Route::get('/journals', [JournalController::class, 'index'])->middleware(['auth', Role::class . ':user']);
+Route::get('/journals/{journal}', [JournalController::class, 'show'])->middleware(['auth', Role::class . ':user']);
 
-Route::patch('/journals/{journal}', [JournalController::class, 'update'])->middleware('auth');
-Route::delete('/journals/{journal}', [JournalController::class, 'destroy'])->name('journals.destroy')->middleware('auth');
+Route::patch('/journals/{journal}', [JournalController::class, 'update'])->middleware(['auth', Role::class . ':user']);
+Route::delete('/journals/{journal}', [JournalController::class, 'destroy'])->name('journals.destroy')->middleware(['auth', Role::class . ':user']);
 
-Route::post('/journals', [JournalController::class, 'store'])->middleware('auth');
+Route::post('/journals', [JournalController::class, 'store'])->middleware(['auth', Role::class . ':user']);
+
+// Admin routes
+Route::get('/therapist-assignments', [AdminTherapistController::class, 'index'])->middleware(['auth', Role::class . ':admin']);
